requirements.txt
aiohttp==3.7.3
alembic==1.5.2
discord.py==1.6.0
flake8==3.8.4
PyDrocsid==1.4.4
PyMySQL==1.0.2
PyYAML==5.4.1
requests==2.25.1
sentry-sdk==0.19.5
SQLAlchemy==1.3.22
